+++ demo/docs-base/.vitepress/theme/index.ts
@@ -8,7 +8,9 @@ import "vitepress-theme-teeker/vp-plus/sidebar.scss";
 import "vitepress-theme-teeker/vp-plus/nav.scss";
 import "vitepress-theme-teeker/vp-plus/aside.scss";
 import "vitepress-theme-teeker/vp-plus/doc-h1.scss";
-// import "vitepress-theme-teeker/vp-plus/container.scss";
+import "vitepress-theme-teeker/vp-plus/nav-blur.scss";
+import "vitepress-theme-teeker/vp-plus/container.scss";
+// import "vitepress-theme-teeker/vp-plus/container-flow.scss";
 // import "vitepress-theme-teeker/vp-plus/blockquote.scss";
 // import "vitepress-theme-teeker/vp-plus/rainbow.scss";