+++ vitepress-theme-teeker/src/components/HomeBanner/src/index.vue
@@ -180,7 +180,10 @@ onUnmounted(() => {
       </p>
     </div>
 
-    <div v-if="features.length && !isBigImgBgStyle" :class="`${ns.e('feature')} flx-wrap-between`">
+    <div
+      v-if="features.length && !isBigImgBgStyle"
+      :class="[ns.e('feature'), 'flx-wrap-between', ns.joinNamespace('wallpaper-outside')]"
+    >
       <div :class="ns.e('feature__item')" v-for="(feature, index) in features" :key="index">
         <a v-if="feature.link" :href="withBase(feature.link)" class="flx-column-center">
           <img v-if="feature.imgUrl" class="feature-img" :src="withBase(feature.imgUrl)" :alt="feature.title" />