+++ vitepress-theme-teeker/src/helper/util.ts
@@ -1,5 +1,5 @@
 /**
- * From vitepress
+ * From Vitepress。Markdown 插件需要用到，因为 Vitepress 的 withBase 方法无法在 Markdown 插件使用
  */
 export const withBase = (base: string, path?: string) => {
   if (!path) return;